Optionally ignore forceLimitPushDown validation
Summary:
Optionally ignore forceLimitPushDown validation

Background:
In the partner insights use case, we shard by partner_id, the query is to get top n pins for a partner group by both image_signature and url. The accurate solution is to get all groups from historical to broker then do top n, but this overwhelms broker for large partners. So it's better to push down limit spec to historicals to better scale. Equivalently, we will get top n pins per day in historicals then in broker get top n of per day top n for all days. Clients are OK with the inaccuracy for large partners.

According to Druid documentation:
"Druid pushes down the limit spec in groupBy queries to the segments on Historicals wherever possible to early prune unnecessary intermediate results and minimize the amount of data transferred to Brokers. By default, this technique is applied only when all fields in the orderBy spec is a subset of the grouping keys. This is because the limitPushDown doesn't guarantee the exact results if the orderBy spec includes any fields that are not in the grouping keys. When all fields in the orderby are part of the grouping key, the Broker will push limit application down to the Historical processes. When the sorting order uses fields that are not in the grouping key, applying this optimization can result in approximate results with unknown accuracy, so this optimization is disabled by default in that case. Enabling this context flag turns on limit push down for limit/orderbys that contain non-grouping key columns." The flag "forceLimitPushDown" can be passed in query context to enable this feature.

Problem:
We are passing SQL to Druid, the group by queries above fails during the SQL planning phase when the "forceLimitPushDown" is turned on. However, when passing Druid native query directly, it works fine. So there's some bug in the SQL planning.

Solution:
The ideal solution is to fix the SQL planning but that requires more effort so a work around here is to add a flag to bypass the unnessary check when we know the query can work fine with limit push down.

Reviewers: #druideng, jgu, jjaffe

Reviewed By: #druideng, jgu, jjaffe

Subscribers: jjaffe, jenkins

JIRA Issue(s): SC-9274

Differential Revision: https://phabricator.pinadmin.com/D485569