Add longs sketch in 0.16
Summary:
Add LongsSketch to support approximate frequencies for items of long type

This is to support another kind of sketch:
https://datasketches.apache.org/docs/Frequency/FrequencySketchesOverview.html

The use case for this is "audience insights" where we want to get the most frequent user interest ids.

There are two implementations, one is ItemsSketch<T> and the other is LongsSketch. This diff will add LongsSketch because we only need long type for now. The upstream LongsSketch implementation from yahoo library doesn't have a direct memory implementation so this diff also adds support to it to better hook into the druid aggregator implementation.

Reviewers: shawncao, dsun, #druideng

Reviewed By: shawncao

Subscribers: jenkins, mleonard, #realtime-analytics

Differential Revision: https://phabricator.pinadmin.com/D549389