Add utilities to Druid docker containers.
Summary: Adding a few utilities to Druid docker containers to make it easier to debug. Adding `less` and `vi` to remove the need to use `cat` and `sed`, which can be painful, and adding `netcat` to allow for better network debugging within a Druid container.

Test Plan: Tested via private build.

Reviewers: #druideng, lucilla

Reviewed By: #druideng, lucilla

Subscribers: jenkins

Differential Revision: https://phabricator.pinadmin.com/D552352