Optimize topN by skipping execution of aggregators that depends on metrics that are not included in the segments
Summary:
Currently, Historical executes all the aggregators of the query on every segment within the interval. With namespacing, this means that conversion aggregators are being executed on onsite segments unnecessarily. Similarly for any datasource that has more than 1 namespace. In a query with a lot of aggregators (e.g. ads query could have up to a few hundreds of aggregator), this severely impacts performance.
The below for loop in topn column selector strategy is the root cause. For large advertiser, the aggregator loop is executed on tens of millions of rows. The solution is to optimize this for-loop to only execute on appropriate aggregators, i.e. aggregators depend on metrics of the segment

```
    while (!cursor.isDone()) {
      final IndexedInts dimValues = selector.getRow();
      for (int i = 0, size = dimValues.size(); i < size; ++i) {
        final int dimIndex = dimValues.get(i);
        final Comparable<?> key = dimensionValueConverter.apply(selector.lookupName(dimIndex));
        Aggregator[] theAggregators = aggregatesStore.get(key);
        if (theAggregators == null) {
          theAggregators = BaseTopNAlgorithm.makeAggregators(
              cursor,
              query.getAggregatorSpecs()
          );
          aggregatesStore.put(key, theAggregators);
        }
        for (Aggregator aggregator : theAggregators) {
          aggregator.aggregate();
        }
      }
      cursor.advance();
      processedRows++;
    }
```
The same strategy can be applied for timeseries and groupby. However topn is a more pressing issue now as we dont have vectorize support for topn

Reviewers: #druideng, lucilla

Reviewed By: #druideng, lucilla

Subscribers: jenkins, mleonard

Differential Revision: https://phabricator.pinadmin.com/D553666

(cherry picked from commit a9613e94dabdedeeef0b782f19f08d130b37a642)