Fixes ordering issue that may cause inaccurate result
Summary:
The previous optimization changes the result order of aggregators by pushing noop aggregators to the end. However, this may cause an issue in the result sequence merge since it relies on ordering of aggregators (not name :( )
This changes basically revert the ordering change. Without pushing the noop aggregators to the end, we have to memorize the positions of those non noop aggregators when evaluate.

Reviewers: #druideng, lucilla

Reviewed By: #druideng, lucilla

Subscribers: jenkins

Differential Revision: https://phabricator.pinadmin.com/D554280