Tune ads-delivery-debugger supervisor spec to account for increased traffic.
Summary: Since late February, the size of data for the ad_delivery_history data source has increased 30-85% per day. This has resulted in larger delays in handoffs between middle managers and historicals and thus higher latency for processing. This diff tunes the supervisor spec to help allay the increase in latency by increasing the number of tasks and decreasing the number of rows before intermediate persists (this would be an ideal use case for setting the max bytes in memory instead of the max rows per segment, but Naehee has recently identified a bug in druid 0.16 that precludes this).

Reviewers: zwang, lucilla, fjaros, #druideng

Reviewed By: lucilla, #druideng

Subscribers: #ads-reporting

Differential Revision: https://phabricator.pinadmin.com/D569692