Adding request b64 encoded to Pepsi Logs
Summary: We want to add the full request to pepsi logs for debugging. Its useful to be able to have the full encoded request to use in combimation with the SendRequestTool in order to replicate and debug issues. We will run compaction every 40 days that removes this field from request logs as this is a large field.

Reviewers: ericnguyen, #druideng

Reviewed By: ericnguyen, #druideng

Subscribers: #ads-reporting

JIRA Issue(s): IM-6473

Differential Revision: https://phabricator.pinadmin.com/D575442