Implement ANY aggregator (#9187) along with code dependencies and  adding custom optimization to only preform ANY aggregations on a single row
Summary: Note this diff contains a custom optimization for ANY aggregator and needs to be treated with care during an update. We optimized ANY aggregations to only be applied to a single row in the response. The custom code for this is in NumericTopNColumnSelectorStrategy and StringTopNColumnSelectorStrategy and BaseTopNAlgorithm

Reviewers: ericnguyen, O1139 Druid

Reviewed By: ericnguyen, O1139 Druid

Subscribers: jgu, jwang, jenkins, mleonard

Differential Revision: https://phabricator.pinadmin.com/D595982

(cherry picked from commit c42dbe226a4bdf4cdbe7aef82e0f6e0f247f8244)