Add in memory bitmap to real time tasks
Summary:
Add in memory bitmap to real time tasks

Motivation: when real time data is ingested through peon tasks, they are first stored in an in memory data structure, a map, when query comes in, every single row in the map is iterated to check if it matches the filter, which is very inefficient. While for segments ingested through batch ingestion, there are bitmap index to pinpoint the candidate rows given the filter which is much more efficient. Currently in order to meet the SLA of some real time use cases, we have to provision a lot of peon tasks and replicas to distribute load which is costly. This change is to add bitmap when data is still in memory to improve query performance and thus lowering the infra needed to support the same or better SLA.

To use:

1) Enable in memory bitmap construction in supervisor spec
```
{
  "type": "kafka",
  "tuningConfig": {
    ...
    "enableInMemoryBitmap": true
    ...
  }
}
```

2) During query time, add a flag in query context to determine if to use the in memory bitmap index (if enabled in step 1)
"useInMemoryBitmapInQuery": true

Test Plan:
Unit test.
Dark read on a prod use case, seeing higher query throughput with much lower number of hosts.

Reviewers: O1139 Druid, jgu

Reviewed By: O1139 Druid, jgu

Subscribers: jenkins, mleonard, #realtime-analytics

Differential Revision: https://phabricator.pinadmin.com/D702366