Update NamedNumberedShardSpec to allow shard spec change
Summary: when changing shardspec from non NamedNumbered to NamedNumbered, such as rolling back from StreamFanOutNamedHashBasedNumberedShardSpec, NamedNumberedShardSpecFactory::create() expected the previous spec has to be namedNumbered or its descedent in order to reuse the partitionName. But that is not necessary since partitionName is already stored in the factory.

Test Plan: tested in dev cluster. while ingestion going on with StreamFanOutNamedHashBasedNumberedShardSpec, update ingestion spec to use NamedNumberedShardSpec.

Reviewers: O1139 Druid, jgu, fjaros

Reviewed By: O1139 Druid, jgu, fjaros

Subscribers: jenkins, #realtime-analytics

Differential Revision: https://phabricator.pinadmin.com/D773581