3f117c972754bc04b92fe4059392b532139757d2 Compress frequency sketch in 64KB blocks instead of compressing individually
Previously each sketch was compressed invidually using lz4 but seems the compression rate is not high possibly because lz4 doesn't work well directly on small data. With additional compression metadata, possibly we are even inflating the data size. According to this issue of lz4 (https://github.com/lz4/lz4/issues/653) for small data (~200B), the compression performance is not that good unless doing some preprocessing of some samples to create a dictionary in advance, which is too complicated and which samples are representative is to be determined. This diff chooses to accumulate more than one sketch to wait for the total byte size to flush accumulates to near 64KB. The compression rate is improved, original compression method generates 500GB audience insights data per day, now it's 90GB per day. Tested with a few top 10 partner_ids, didn't see obvious latency increase.